.net core, network queues rework
Updated the Bridge Application to .NET Core and converted from file item transfer to network queues. Threaded everything to make the whole application run better. Need to update how item mapping is done to make it more readable.

Factorio version is now v1.0 of official release. Updated Factorio to now use only interface to send and receive items when instructed to by the bridge. This should help transfer errors. Added  the ability to send and receive fuilds and power. Added UI's to change priorities and apply filters to the chests and tanks.

Updated forge version for 1.12.2 to latest. Updated minecraft to use sockets now to connect to the bridge for the queues. Items can be transfer successfully. UI is still WIP for filtering and priorities and fluid and power still need to be worked on.